subscription_script/pyproject.toml
[project]
name = "subscription-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ReySKR", email = "stephanr.tin23@student.dhbw-heidenheim.de" }
]
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]

[project.scripts]
subscription-script = "subscription_script:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]
